Consolidate region error reporting in rustc_infer
